moved pagedresults to pagination package
